A ton of bug fixes
Fixed commands to get the log name from the export data. Fixed the actual logging not being done correctly and in the correct folder. Set logs to delete after 14 days. Fixed the countdown database bug. Actually made the bod sync the tables on startup so that on the first time run all the tables are created before they are needed. No more database errors.